# docs/v1/swagger.yaml
basePath: /api/web-api-gateway/v1
definitions:
...
  domain.ErrorResponse:
    properties:
      code:
        type: integer
      data: {}
      is_error:
        type: boolean
      message:
        type: string
    type: object
  domain.InputSupplier:
    properties:
      actual_address:
        description: Фактический адрес поставщика
        example: Фактический адрес поставщика
        type: string
      balance:
        description: Баланс по поставщику
        example: 100
        type: number
      bank_details:
        description: Банковские реквизиты поставщика
        example: Банковские реквизиты поставщика
        type: string
      comments:
        description: Комментарии
        example: Комментарии
        type: string
      contact_person:
        description: Контактное лицо у поставщика
        example: Контактное лицо у поставщика
        maxLength: 140
        type: string
      contract_number:
        description: Номер и дата договора с поставщиком
        example: Номер и дата договора с поставщиком
        type: string
      country:
        description: Страна поставщика
        example: Страна поставщика
        type: string
      email:
        description: Электронная почта поставщика
        example: Электронная почта поставщика
        type: string
      files:
        description: Ссылки на файлы или документы
        example: Ссылки на файлы или документы
        type: string
      is_active:
        description: Статус активности поставщика (активен/неактивен)
        example: true
        type: boolean
      legal_address:
        description: Юридический адрес поставщика
        example: Юридический адрес поставщика
        type: string
      name:
        description: Наименование поставщика
        example: ООО Название поставщика
        maxLength: 140
        minLength: 5
        type: string
      other_fields:
        additionalProperties: true
        description: Дополнительные пользовательские поля
        type: object
      payment_terms:
        description: Условия оплаты по контракту
        example: Условия оплаты по контракту
        type: string
      phone:
        description: Телефон поставщика
        example: Телефон поставщика
        minLength: 5
        type: string
      product_categories:
        description: Категории товаров, поставляемых поставщиком
        example: Категории товаров, поставляемых поставщиком
        type: string
      product_types:
        description: Количество типов товаров от поставщика
        example: 100
        type: integer
      purchase_amount:
        description: Общая сумма закупок у поставщика
        example: 100
        type: number
      region:
        description: Регион или штат поставщика
        example: Регион или штат поставщика
        type: string
      registration_date:
        description: Дата регистрации поставщика
        example: "2022-01-01T00:00:00Z"
        type: string
      tax_id:
        description: Идентификационный номер налогоплательщика (ИНН)
        example: Идентификационный номер налогоплательщика (ИНН)
        type: string
      warehouse_address:
        description: Адрес склада поставщика
        example: Адрес склада поставщика
        type: string
      website:
        description: Сайт поставщика
        example: Сайт поставщика
        type: string
    required:
    - actual_address
    - balance
    - bank_details
    - contact_person
    - contract_number
    - legal_address
    - name
    - phone
    - warehouse_address
    type: object
...
  domain.SuccessResponse:
    properties:
      data: {}
      is_error:
        type: boolean
      total_count:
        type: integer
    type: object
...
  domain.UpdateSupplier:
    properties:
      actual_address:
        description: Фактический адрес поставщика
        example: Фактический адрес поставщика
        type: string
      balance:
        description: Баланс по поставщику
        example: 100
        type: number
      bank_details:
        description: Банковские реквизиты поставщика
        example: Банковские реквизиты поставщика
        type: string
      comments:
        description: Комментарии
        example: Комментарии
        type: string
      contact_person:
        description: Контактное лицо у поставщика
        example: Контактное лицо у поставщика
        maxLength: 140
        type: string
      contract_number:
        description: Номер и дата договора с поставщиком
        example: Номер и дата договора с поставщиком
        type: string
      country:
        description: Страна поставщика
        example: Страна поставщика
        type: string
      email:
        description: Электронная почта поставщика
        example: Электронная почта поставщика
        type: string
      files:
        description: Ссылки на файлы или документы
        example: Ссылки на файлы или документы
        type: string
      is_active:
        description: Статус активности поставщика (активен/неактивен)
        example: true
        type: boolean
      legal_address:
        description: Юридический адрес поставщика
        example: Юридический адрес поставщика
        type: string
      name:
        description: Наименование поставщика
        example: ООО Название поставщика
        maxLength: 140
        minLength: 5
        type: string
      other_fields:
        additionalProperties: true
        description: Дополнительные пользовательские поля
        type: object
      payment_terms:
        description: Условия оплаты по контракту
        example: Условия оплаты по контракту
        type: string
      phone:
        description: Телефон поставщика
        example: Телефон поставщика
        minLength: 5
        type: string
      product_categories:
        description: Категории товаров, поставляемых поставщиком
        example: Категории товаров, поставляемых поставщиком
        type: string
      product_types:
        description: Количество типов товаров от поставщика
        example: 100
        type: integer
      purchase_amount:
        description: Общая сумма закупок у поставщика
        example: 100
        type: number
      region:
        description: Регион или штат поставщика
        example: Регион или штат поставщика
        type: string
      registration_date:
        description: Дата регистрации поставщика
        example: "2022-01-01T00:00:00Z"
        type: string
      tax_id:
        description: Идентификационный номер налогоплательщика (ИНН)
        example: Идентификационный номер налогоплательщика (ИНН)
        type: string
      warehouse_address:
        description: Адрес склада поставщика
        example: Адрес склада поставщика
        type: string
      website:
        description: Сайт поставщика
        example: Сайт поставщика
        type: string
    required:
    - actual_address
    - balance
    - bank_details
    - contact_person
    - contract_number
    - legal_address
    - name
    - phone
    - warehouse_address
    type: object
...
host: 91.243.71.100:8080
info:
  contact:
    email: ru.system.ru@gmail.com
    name: ru.system.ru@gmail.com
  description: API gateway
  title: Web api gateway API
  version: "1.0"
paths:
...
  /supplier:
    get:
      consumes:
      - application/json
      description: Получение всех поставщиков компании
      operationId: get-suppliers
      parameters:
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort
        required: true
        type: string
      - default: name
        description: Field to sort by
        enum:
        - id
        - name
        - legal_address
        - actual_address
        - warehouse_address
        - contact_person
        - phone
        - email
        - website
        - contract_number
        - product_categories
        - purchase_amount
        - balance
        - product_types
        - country
        - region
        - tax_id
        - registration_date
        - is_active
        in: query
        name: sort_field
        required: true
        type: string
      - description: limit query param
        in: query
        name: limit
        required: true
        type: integer
      - description: offset query param
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get suppliers
      tags:
      - supplier
    post:
      consumes:
      - application/json
      description: Создание поставщика
      operationId: create-supplier
      parameters:
      - description: Необходимо указать данные поставщика.
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.InputSupplier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create supplier
      tags:
      - supplier
  /supplier/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление поставщика своей компании
      operationId: delete-supplier
      parameters:
      - description: ID поставщика
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete supplier
      tags:
      - supplier
    get:
      consumes:
      - application/json
      description: Получение поставщика по id
      operationId: get-supplier
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get supplier by id
      tags:
      - supplier
    put:
      consumes:
      - application/json
      description: Обновление поставщика
      operationId: update-supplier
      parameters:
      - description: ID поставщика
        in: path
        name: id
        required: true
        type: integer
      - description: Необходимо указать данные поставщика.
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateSupplier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update supplier
      tags:
      - supplier
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"